Send SMTPUTF8 if the message needs it and the server supports it.
Before this commit, Envelope would throw InvalidArgumentException when a
unicode sender address was used. Now, that error is thrown slightly later,
is thrown for recipient addresses as well, but is not thrown if the
next-hop server supports SMTPUTF8.

As a side effect, transports that use JSON APIs to ESPs can also use
unicode addresses if the ESP supports that (many do, many don't).